Extract multi signature assets
