+++ packages/contracts/source/contracts/crypto/index.ts
@@ -5,6 +5,7 @@ export * from "./crypto.js";
 export * from "./enums.js";
 export * from "./hash.js";
 export * from "./identities.js";
+export * from "./legacy.js";
 export * from "./messages.js";
 export * from "./networks.js";
 export * from "./signatures.js";
