+++ packages/contracts/source/contracts/crypto/legacy.ts
@@ -1,0 +1,10 @@
+export interface MultiSignatureLegacyAsset {
+	min: number;
+	lifetime: number;
+	keysgroup: string[];
+}
+
+export interface MultiSignatureAsset {
+	min: number;
+	publicKeys: string[];
+}
